requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
httpx==0.27.0
faster-whisper==1.0.1
numpy==1.26.4
aiofiles==23.2.1
python-dateutil==2.9.0
elevenlabs==1.2.2

# LangChain
langchain==0.2.6
langchain-google-genai==1.0.6
langchain-core==0.2.10
google-generativeai==0.7.2
